feat: renamed EmailDIsplay with EmailDisplay.tsx
