+++ frontend/libs/erxes-ui/src/modules/display/components/EmailDisplay.tsx
