+++ frontend/libs/erxes-ui/src/modules/display/index.ts
@@ -2,5 +2,5 @@ export * from './components/CurrencyDisplay';
 export * from './components/EllipsisDisplay';
 export * from './components/RelativeDateDisplay';
 export * from './components/PhoneDisplay';
-export * from './components/EmailDIsplay';
+export * from './components/EmailDisplay';
 export * from './components/CurrencyDisplay';
